New matrix workflow template
This new workflow uses the `matrix_from_metadata` script from puppet_litmus
to calculate the required GCP image ids and create an appropriate workflow
matrix.

It also uses updated versions of actions and cleans up the cache.

Note that there is no slack notifications on the PR
workflow as that requires access to secrets that are
not available on workflows running on forks. Given that
PRs also can be very noisy, I don't consider that a big
loss.